Added optional scaling of wpfbt coefficients. Backend rewrittent to a C99 complex.h number format.

+++ blockproc/java/net/sourceforge/ltfat/Utils.java
@@ -171,7 +171,7 @@ public class Utils {
     {
        int height = in.length;
        for(int yy=0;yy<height;yy++){
-             in[yy] = (float) (20*Math.log10(in[yy]+1e-10));
+             in[yy] = 20.0*Math.log10(in[yy]+1e-10);
        }
     }
